route restriction

+++ src/modules/jobRequirement/routes/jobRequirementRoutes.js
@@ -2,7 +2,7 @@ const express = require('express');
 const jobRequirementRouter = express.Router();
 const jobRequirementCtrl = require('../controller/jobRequirementController');
 
-let router = function (connection) {
+let router = function (role) {
     jobRequirementRouter.route('/')
         .get(function (req, res) {
             jobRequirementCtrl.getAllJobRequirment()
@@ -21,20 +21,28 @@ let router = function (connection) {
                 });
         })
         .post(function (req, res) {
-            jobRequirementCtrl.createJobSkills(req.body)
-                .then((result) => {
-                    res.json({
-                        success: true,
-                        data: result
+            if (role > 1) {
+                jobRequirementCtrl.createJobSkills(req.body)
+                    .then((result) => {
+                        res.json({
+                            success: true,
+                            data: result
+                        });
+                    })
+                    .catch(function (error) {
+                        res.status(400);
+                        res.json({
+                            success: false,
+                            data: error.toString()
+                        });
                     });
-                })
-                .catch(function (error) {
-                    res.status(400);
-                    res.json({
-                        success: false,
-                        data: error.toString()
-                    });
+            } else {
+                res.status(403);
+                res.json({
+                    success: false,
+                    data: "Forbidden"
                 });
+            }
         });
 
     jobRequirementRouter.route('/:id')
@@ -55,20 +63,28 @@ let router = function (connection) {
                 });
         })
         .delete(function (req, res) {
-            jobRequirementCtrl.deleteJobSkill(req.params.id)
-                .then((result) => {
-                    res.json({
-                        success: true,
-                        data: result
+            if (role > 1) {
+                jobRequirementCtrl.deleteJobSkill(req.params.id)
+                    .then((result) => {
+                        res.json({
+                            success: true,
+                            data: result
+                        });
+                    })
+                    .catch(function (error) {
+                        res.status(400);
+                        res.json({
+                            success: false,
+                            data: error.toString()
+                        });
                     });
-                })
-                .catch(function (error) {
-                    res.status(400);
-                    res.json({
-                        success: false,
-                        data: error.toString()
-                    });
+            } else {
+                res.status(403);
+                res.json({
+                    success: false,
+                    data: "Forbidden"
                 });
+            }
         });
 
     return jobRequirementRouter;
